Move `WaitingSender` to its own module
