+++ src/inbox.rs
@@ -4,19 +4,18 @@
 pub mod rx;
 pub mod tx;
 mod waiting_receiver;
+mod waiting_sender;
 
 use std::cmp::Ordering;
 use std::collections::{BinaryHeap, VecDeque};
-use std::future::Future;
-use std::pin::Pin;
 use std::sync::atomic::AtomicUsize;
 use std::sync::{atomic, Arc, Mutex, Weak};
-use std::task::{Context, Poll, Waker};
 use std::{cmp, mem};
 
 use event_listener::{Event, EventListener};
 pub use rx::Receiver;
 pub use tx::Sender;
+pub use waiting_sender::WaitingSender;
 
 use crate::envelope::{BroadcastEnvelope, MessageEnvelope, Shutdown};
 use crate::inbox::tx::TxStrong;
@@ -538,76 +537,6 @@ impl<A> From<Box<dyn MessageEnvelope<Actor = A>>> for ActorMessage<A> {
 impl<A> From<Arc<dyn BroadcastEnvelope<Actor = A>>> for ActorMessage<A> {
     fn from(msg: Arc<dyn BroadcastEnvelope<Actor = A>>) -> Self {
         ActorMessage::ToAllActors(msg)
-    }
-}
-
-pub enum WaitingSender<M> {
-    Active { waker: Option<Waker>, message: M },
-    Delivered,
-    Closed,
-}
-
-impl<M> WaitingSender<M> {
-    pub fn new(message: M) -> Arc<Spinlock<Self>> {
-        let sender = WaitingSender::Active {
-            waker: None,
-            message,
-        };
-        Arc::new(Spinlock::new(sender))
-    }
-
-    pub fn peek(&self) -> &M {
-        match self {
-            WaitingSender::Active { message, .. } => message,
-            _ => panic!("WaitingSender should have message"),
-        }
-    }
-
-    pub fn fulfill(&mut self) -> M {
-        match mem::replace(self, Self::Delivered) {
-            WaitingSender::Active { mut waker, message } => {
-                if let Some(waker) = waker.take() {
-                    waker.wake();
-                }
-
-                message
-            }
-            WaitingSender::Delivered | WaitingSender::Closed => {
-                panic!("WaitingSender is already fulfilled or closed")
-            }
-        }
-    }
-
-    /// Mark this [`WaitingSender`] as closed.
-    ///
-    /// Should be called when the last [`Receiver`](crate::inbox::Receiver) goes away.
-    pub fn set_closed(&mut self) {
-        if let WaitingSender::Active {
-            waker: Some(waker), ..
-        } = mem::replace(self, Self::Closed)
-        {
-            waker.wake();
-        }
-    }
-}
-
-impl<M> Future for WaitingSender<M>
-where
-    M: Unpin,
-{
-    type Output = Result<(), Error>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let this = self.get_mut();
-
-        match this {
-            WaitingSender::Active { waker, .. } => {
-                *waker = Some(cx.waker().clone());
-                Poll::Pending
-            }
-            WaitingSender::Delivered => Poll::Ready(Ok(())),
-            WaitingSender::Closed => Poll::Ready(Err(Error::Disconnected)),
-        }
     }
 }
